Reorganizar importaciones para utilizar componentes compartidos en app-header y app-sidebar
