+++ resources/js/layouts/app/app-sidebar-layout.tsx
@@ -1,7 +1,7 @@
-import { AppContent } from '@/components/app-content';
-import { AppShell } from '@/components/app-shell';
-import { AppSidebar } from '@/components/app-sidebar';
-import { AppSidebarHeader } from '@/components/app-sidebar-header';
+import { AppContent } from '@/components/shared/app-content';
+import { AppShell } from '@/components/shared/app-shell';
+import { AppSidebar } from '@/components/shared/app-sidebar';
+import { AppSidebarHeader } from '@/components/shared/app-sidebar-header';
 import { type BreadcrumbItem } from '@/types';
 import { type PropsWithChildren } from 'react';
